:robot: automated commit of 1-array_iterator.c

+++ function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int temp = 0;
+	size_t temp = 0;
 
 	while (temp < size)
 	{
